Implementation of dynamic List to add vehicles and delete vehicles
it was added, that there is the possibility to add multiple vehicles with different names and different ipadresses to the app. That allows the user to change betweeen vehicles or use the car in different networts as long as both, the vehicle and the phone are in the same network.